requirements2.txt
numpydoc==0.8.0
pandas==0.24.2
scikit-image==0.14.2
scikit-learn==0.20.3
scikit-surprise==1.0.6
scipy==1.2.1
